refactor(features): 배럴 public surface 정리 — audit-log 구체 repository 누출 제거 + 미사용 seller/user 배럴 삭제
