+++ src/features/conversation/index.ts
@@ -1,2 +1,3 @@
+// cross-feature 공개 API. 단일 구현 repo라 토큰/인터페이스 없이 구체 클래스로 주입(의도적).
 export { ConversationModule } from '@/features/conversation/conversation.module';
 export { ConversationRepository } from '@/features/conversation/repositories/conversation.repository';
