+++ src/features/product/index.ts
@@ -1,2 +1,3 @@
+// cross-feature 공개 API. 단일 구현 repo라 토큰/인터페이스 없이 구체 클래스로 주입(의도적).
 export { ProductModule } from '@/features/product/product.module';
 export { ProductRepository } from '@/features/product/repositories/product.repository';
