--- src/features/seller/index.ts
@@ -1,2 +1,0 @@
-export { SellerModule } from '@/features/seller/seller.module';
-export { SellerRepository } from '@/features/seller/repositories/seller.repository';
