+++ src/features/seller/services/seller-audit.service.spec.ts
@@ -1,7 +1,8 @@
 import { BadRequestException } from '@nestjs/common';
 import { Prisma, type PrismaClient } from '@prisma/client';
 
-import { AUDIT_LOG_REPOSITORY, AuditLogRepository } from '@/features/audit-log';
+import { AUDIT_LOG_REPOSITORY } from '@/features/audit-log';
+import { AuditLogRepository } from '@/features/audit-log/repositories/audit-log.repository';
 import { SellerRepository } from '@/features/seller/repositories/seller.repository';
 import { SellerAuditService } from '@/features/seller/services/seller-audit.service';
 import { SellerFaqService } from '@/features/seller/services/seller-faq.service';
