--- src/features/user/index.ts
@@ -1,2 +1,0 @@
-export { UserModule } from '@/features/user/user.module';
-export { UserRepository } from '@/features/user/repositories/user.repository';
